+++ src_cpp/Lidar.hpp
@@ -14,7 +14,7 @@ public:
     // Escolha: size = 5. Com a janela cíclica de 100ms, uma janela de tamanho 5
     // dá um atraso de meio segundo, filtrando picos isolados do sensor.
     LidarFilter(int size = 5);
-    float calcular(int nova_leitura);
+    float calcular(float nova_leitura);
     float get_ultima_media();
 };
